requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic[email]==2.3.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.0.1
motor==3.3.1
motor==3.3.1
